+++ src/components/repair/request-repair/index.js
@@ -1,0 +1,6 @@
+export { default as StepPayer } from './StepPayer';
+export { default as StepForm } from './StepForm';
+export { default as StepReview } from './StepReview';
+export { default as StepDone } from './StepDone';
+export * from './constants';
+export * from './styles';
